refactor: introduce vendor-specific types and simplify vendor configuration
This commit introduces vendor-specific types and simplifies the vendor configuration.
