+++ packages/common/src/configuration/vendor.ts
@@ -14,22 +14,3 @@ export const VendorConfig = z.object({
 });
 
 export type VendorConfig = z.infer<typeof VendorConfig>;
-
-export const PochiVendorConfig = VendorConfig.extend({
-  credentials: z.object({
-    token: z.string(),
-  }),
-});
-
-export type PochiVendorConfig = z.infer<typeof PochiVendorConfig>;
-
-export const GeminiCliVendorConfig = VendorConfig.extend({
-  credentials: z.object({
-    accessToken: z.string(),
-    refreshToken: z.string(),
-    expiresAt: z.number(),
-    project: z.string(),
-  }),
-});
-
-export type GeminiCliVendorConfig = z.infer<typeof GeminiCliVendorConfig>;
