--- packages/vendor-pochi/src/constants.ts
@@ -1,1 +1,0 @@
-export const VendorId = "pochi";
