+++ packages/vendor-pochi/src/types.ts
@@ -1,0 +1,5 @@
+export const VendorId = "pochi";
+
+export type PochiCredentials = {
+  token: string;
+};
