+++ packages/vscode/src/lib/token-storage.ts
@@ -1,4 +1,5 @@
-import { pochiConfig, updatePochiConfig } from "@getpochi/common/configuration";
+import { updatePochiConfig } from "@getpochi/common/configuration";
+import { getPochiCredentials } from "@getpochi/vendor-pochi";
 import { computed } from "@preact/signals-core";
 import { injectable, singleton } from "tsyringe";
 import type * as vscode from "vscode";
@@ -6,7 +7,7 @@ import type * as vscode from "vscode";
 @injectable()
 @singleton()
 export class TokenStorage implements vscode.Disposable {
-  token = computed(() => pochiConfig.value.vendors?.pochi?.credentials?.token);
+  token = computed(() => getPochiCredentials()?.token);
   dispose: () => void = () => {};
 
   setToken(token: string | undefined) {
